requirements.txt
mp-api
pymatgen
google-genai
python-dotenv
